+++ samples/Texture2DArray/main.cpp
@@ -106,7 +106,7 @@ float4 PSMain(float4 position : SV_Position, float2 uv : TEXCOORD0) : SV_Target0
 		desc.fragmentShader.source = pixelShader;
 		desc.fragmentShader.entryPoint = "PSMain";
 		desc.fragmentShader.profile = "ps_6_6";
-		desc.colorFormat = DXGI_FORMAT_R8G8B8A8_UNORM;
+		desc.color[ 0 ].format = DXGI_FORMAT_R8G8B8A8_UNORM;
 		desc.depthFormat = DXGI_FORMAT_UNKNOWN;
 		desc.depthStencilState.DepthEnable = FALSE;
 		desc.depthStencilState.StencilEnable = FALSE;