+++ samples/TextureSamplers/main.cpp
@@ -98,7 +98,7 @@ float4 PSMain(PixelInput input) : SV_Target0
 		desc.fragmentShader.entryPoint = "PSMain";
 		desc.fragmentShader.profile = "ps_6_6";
 		desc.fragmentShader.sourceName = "TextureSamplersPS";
-		desc.colorFormat = DXGI_FORMAT_R8G8B8A8_UNORM;
+		desc.color[ 0 ].format = DXGI_FORMAT_R8G8B8A8_UNORM;
 		desc.depthFormat = DXGI_FORMAT_UNKNOWN;
 		desc.depthStencilState.DepthEnable = FALSE;
 		desc.depthStencilState.StencilEnable = FALSE;