+++ samples/ZFighting/main.cpp
@@ -102,7 +102,7 @@ float4 PSMain(float4 position : SV_Position, float4 color : COLOR0) : SV_Target0
 		desc.fragmentShader.entryPoint = "PSMain";
 		desc.fragmentShader.profile = "ps_6_6";
 		desc.fragmentShader.sourceName = "ZFightingPS";
-		desc.colorFormat = DXGI_FORMAT_R8G8B8A8_UNORM;
+		desc.color[ 0 ].format = DXGI_FORMAT_R8G8B8A8_UNORM;
 		desc.depthFormat = DXGI_FORMAT_D32_FLOAT;
 		desc.sampleCount = sampleCount;
 		desc.rasterizerState.CullMode = D3D12_CULL_MODE_NONE;